float

+++ calculator.py
@@ -1,5 +1,5 @@
-x = int(input("what is x?"))
-y = int(input("what is y?"))
+x = float(input("what is x?"))
+y = float(input("what is y?"))
 
 z = x+y
